refactors library, adds breadcrumbs to flat course
