+++ src/lib/index.js
@@ -1,3 +1,13 @@
 export * from "./totalTime";
 export * from "./topicManager";
-export * from "./toJSON";
+
+export const toJSON = (res) => res.json();
+
+export const toText = (res) => res.text();
+
+export const throwIt = (msg) => (error) => {
+  if (msg) {
+    console.error(msg);
+  }
+  throw error;
+};
