--- src/lib/toJSON.js
@@ -1,10 +1,0 @@
-export const toJSON = (res) => res.json();
-
-export const toText = (res) => res.text();
-
-export const throwIt = (msg) => (error) => {
-  if (msg) {
-    console.error(msg);
-  }
-  throw error;
-};
