--- src/lib/toJSON.test.js
@@ -1,9 +1,0 @@
-import { toJSON } from "./toJSON";
-
-describe("toJSON", () => {
-  it("calls .json() on the response", () => {
-    const fn = jest.fn();
-    toJSON({ json: fn });
-    expect(fn).toBeCalledTimes(1);
-  });
-});
