+++ src/lib/topicManager.test.data.js
@@ -28,6 +28,7 @@ export const flatSampleCourse = [
     id: "sample-course",
     title: "sample course",
     agenda: ["chapter-one"],
+    breadcrumbs: [],
     time: {
       est: 10,
     },
@@ -36,6 +37,7 @@ export const flatSampleCourse = [
     id: "chapter-one",
     title: "Chapter One",
     agenda: ["step-one"],
+    breadcrumbs: ["sample-course"],
     time: {
       est: 10,
     },
@@ -44,6 +46,7 @@ export const flatSampleCourse = [
     id: "step-one",
     title: "step one",
     agenda: ["task-a", "task-b"],
+    breadcrumbs: ["sample-course", "chapter-one"],
     time: {
       est: 10,
     },
@@ -51,6 +54,7 @@ export const flatSampleCourse = [
   {
     id: "task-a",
     title: "task A",
+    breadcrumbs: ["sample-course", "chapter-one", "step-one"],
     time: {
       length: 5,
       est: 5,
@@ -60,6 +64,7 @@ export const flatSampleCourse = [
     id: "task-b",
     title: "task B",
     agenda: ["challenge-1", "challenge-2"],
+    breadcrumbs: ["sample-course", "chapter-one", "step-one"],
     time: {
       length: 5,
       est: 5,
@@ -68,6 +73,7 @@ export const flatSampleCourse = [
   {
     id: "challenge-1",
     title: "Challenge 1",
+    breadcrumbs: ["sample-course", "chapter-one", "step-one", "task-b"],
     time: {
       est: 2.5,
     },
@@ -75,6 +81,7 @@ export const flatSampleCourse = [
   {
     id: "challenge-2",
     title: "Challenge 2",
+    breadcrumbs: ["sample-course", "chapter-one", "step-one", "task-b"],
     time: {
       est: 2.5,
     },
